Core: overhaul `Filter_File_Handle`
The handle contained unused or unnecessary functions, exposed its
internal string-stream and had some unnecessarily confusing naming. In
order not to introduce a huge amount of changes throughout the entire
codebase, the general idea is kept the same:
 - the handle throws if a file cannot be opened
 - the user manipulates the file stream by calling searching functions,
   like `Find` or `Read_String` and subsequently streams out of the
   handle, `file_handle >> variable`. This allows the user to flexibly
   control what happens when a given keyword was found.

The code should now be a little easier to understand and further improve
in the future. Aside from the above, more things are now `const` and
properly initialized upon construction.

Also made a few minor quality improvements in a few places.