Quantities: Get_Magnetization now respects mu_s
Previously, the values of mu_s were ignored, so that the computed quantity was only the average of the spin directions. Unfortunately computing the magnetization now requires a temporary field of Vector3 (especially when using CUDA), which has negative performance implications. Therefore the new function `Quantity_Get_Average_Spin` has been introduced which is now used to display the average of the spin direction in the QT GUI. Also the magnetization has been removed from the `history` of Method_LLG. Some unit tests had to be adjusted, because they were relying on the old (wrong) behaviour of Quantity_Get_Magnetization