Quantities: Added topological charge to unit test
Also changed the `api.cfg` input file in the quantities unit test to be more general and use an additional basis atom instead of a monoatomic basis. The quantities unit test now uses the `api.cfg` input file instead of `fd_neighbours.cfg`, because the system size should be actually be able to fit an entire skyrmion.