Hamiltonian_Heisenberg: Fixed race condition in DDI
In the pointwise multiplication the parallelization was collapsing one loop too many, and thus parallelized the summation of the inter-sublattice contributions. This lead to a racecondition when n_cell_atoms > 1. The effects of this racecondition were only really seen when using a very high number of threads compared to the system size.