refactor(modular Hamiltonian): connect Interaction::Gaussian data
Introduces the data structure for storing interactions. It is of type:
`std::vector<std::unique_ptr<Engine::Interaction::ABC>>`
This vector marks the `Hamiltonian` as the owner of the interactions and
necessitates the creation of constructors according to the rule of five.

For setting and getting the interactions a set of templated member
functions is used which can also be used at run time given the name of
the interaction.

The Gaussian interaction is not usable with this commit.