refactor(`TableParser`): centralize common logic for parsing data tables
flexible parsing of table data whose column order isn't known at compile
time has lead to a lot of code duplication. The `TableParser` class
fixes that by putting the logic for reading the rows of this type of
table in a single place. The `parse` method allows specifying how the
data from each row should be transformed before it is stored.
The only draw back is that a second post processing step is needed once
all of the data has been read.