feature(Biaxial Anisotropy): introduce more complex anisotropy term
This implementation of the anisotropy supports complex realisations of
an anisotropy energy landscape. The two anisotropy axes can be specified
independently for each site in the unit cell as well as a list of
coefficients with a set of three exponents for the angular dependencies.
For details see `core/docs/Hamiltonian.md`.