enable templates for `Spin_System` and `Spin_System_Chain`
Turn `Spin_System` and `Spin_System_Chain` into templates to prepare for
the introduction of a second Hamiltonian that needs a different
representation of the state of the system.

The State struct is used to communicate the necessary types to the
different call sites, because the C-API forbids the use of Templates in
some places. This makes the coupling to these types explicit.