restructure files to make room for a second Hamiltonian
Method specializations and Solvers are all coupled to the Hamiltonian,
so I need a second set of them (for now) to implement a second Hamiltonian
with the associated Methods specializations and Solvers.

The Method interface is left in the common directory, because it is
generic enough to serve as a common interface for both implementations.