make the Geometry object private to the Hamiltonian
The internal state of the Hamiltonian depends on the state of the
Geometry object it is tied to. To ensure that the state of Hamiltonian
remains valid, we have to give the Hamiltonian control over the Geometry
object.

This commit breaks the direct write access to parts of the Geometry. We
might consider adding that back in and putting the responsibility of
triggering the update of the internal state of the Hamiltonian into the
hands of the API user. But for now this solution is more straight
forward to implement.