decouple `io/Datawriter` from `Data::Spin_System`
Factor out the energy and magnetization data that are used for caches
from the templated `Data::Spin_System` class. This allows the
`IO::*_Image_Energy` functions to operate on the `System_Energy` structs
instead.

This also reintroduces the `energy_contributions_per_spin` variable as a
cached value that is now part of the `System_Energy` struct.