enhancement(API): less greedy resize of spin vectors trough the API
Previously the vectors containing atom local quantities were updated
whenever there was any change to the geometry. That means that any
change to the geometry (like setting `mu_s`) invalidates the pointer
to these quantities, like the pointer returned by
`Configuration_Get_Spin_Directions()`.
For `mu_s` this is unexpected, because it requires deep knowledge of
implementation details.
This commit therefore adds a check whether a resize is necessary, by
checking whether the size described by two geometry objects is the
same and only does a resize if they differ. This should result in more
intuitive behaviour when it comes to changing quantities that are
related with the geometry through our implementation.