refactor(toml_config): unified vector parsing
Unifies the implementation and config file interface for parsing vectors.
If it is a 'true' vector, like the temperature gradient or the external
field each vector can be specified either as an array, or as a table
with fields 'direction' and 'magnitude'. The vector is then internally
stored in terms of these two quantities.

If it is a quantity that has an associated normalized vector and a
magnitude, like the uniaxial anisotropy the fields are 'normal' and
'magnitude'. This time also both have to be provided, because there is
ambiguity in how the magnitude might be determined from the norm of the
provided vector.

This skips the `stt_magnitude` and `stt_polarization_direction` fields,
because those should be adapted to be more understandable.