Move `boundary_conditions` inside `Data::Geometry`
The `boundary_conditions` have been related more closely to the
`Geometry` than the `Hamiltonian` for some time now. This move also
reflects the correspondence in the code, which also allows moving it to
the Geometry section in the Config and the API.

TODO: Move the boundary conditions to the Geometry widgets.